# swagger/AZURE-ETL-API.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:4000/",
      "description": "ETL API for Azure Face API",
      "variables": {}
    }
  ],
  "info": {
    "version": "0.0.1",
    "title": "AZURE FACE ETL API",
    "description": "",
    "termsOfService": "",
    "contact": {
      "name": "Prashant Pandey"
    },
    "license": {
      "name": ""
    }
  },
  "paths": {
    "/v1/face": {
      "post": {
        "summary": "Get attributes and features of faces in the uploaded image",
        "description": "API can be used to detect human faces in an image and returns the face rectangles and faceIds, and optionally with location, features, landmark, emotions, characterstics, and noises.",
        "operationId": "FaceDetection",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "example": [
                    {
                      "faceId": "a261eba4-3d81-40b9-80de-863513b2cffd",
                      "faceRectangle": {
                        "top": 72,
                        "left": 261,
                        "width": 130,
                        "height": 130
                      },
                      "faceLandmarks": {
                        "pupilLeft": {
                          "x": 299.5,
                          "y": 107.1
                        },
                        "pupilRight": {
                          "x": 352.5,
                          "y": 106.2
                        },
                        "noseTip": {
                          "x": 314.1,
                          "y": 143.3
                        },
                        "mouthLeft": {
                          "x": 306.9,
                          "y": 169
                        },
                        "mouthRight": {
                          "x": 359.3,
                          "y": 168.2
                        },
                        "eyebrowLeftOuter": {
                          "x": 283,
                          "y": 93.7
                        },
                        "eyebrowLeftInner": {
                          "x": 313.5,
                          "y": 96.1
                        },
                        "eyeLeftOuter": {
                          "x": 293.4,
                          "y": 108.4
                        },
                        "eyeLeftTop": {
                          "x": 302.9,
                          "y": 103.3
                        },
                        "eyeLeftBottom": {
                          "x": 302.7,
                          "y": 111
                        },
                        "eyeLeftInner": {
                          "x": 311.5,
                          "y": 108.2
                        },
                        "eyebrowRightInner": {
                          "x": 335.9,
                          "y": 93.6
                        },
                        "eyebrowRightOuter": {
                          "x": 380.4,
                          "y": 98.1
                        },
                        "eyeRightInner": {
                          "x": 346.6,
                          "y": 108
                        },
                        "eyeRightTop": {
                          "x": 357.6,
                          "y": 102.1
                        },
                        "eyeRightBottom": {
                          "x": 356.5,
                          "y": 110.8
                        },
                        "eyeRightOuter": {
                          "x": 366.3,
                          "y": 108.7
                        },
                        "noseRootLeft": {
                          "x": 316.6,
                          "y": 108.8
                        },
                        "noseRootRight": {
                          "x": 335.9,
                          "y": 109.4
                        },
                        "noseLeftAlarTop": {
                          "x": 309.4,
                          "y": 130
                        },
                        "noseRightAlarTop": {
                          "x": 336.3,
                          "y": 132
                        },
                        "noseLeftAlarOutTip": {
                          "x": 301.8,
                          "y": 140.6
                        },
                        "noseRightAlarOutTip": {
                          "x": 343.2,
                          "y": 141.8
                        },
                        "upperLipTop": {
                          "x": 322.7,
                          "y": 162.2
                        },
                        "upperLipBottom": {
                          "x": 323.3,
                          "y": 167.2
                        },
                        "underLipTop": {
                          "x": 324,
                          "y": 180.9
                        },
                        "underLipBottom": {
                          "x": 324.9,
                          "y": 188.5
                        }
                      },
                      "faceAttributes": {
                        "smile": 1,
                        "headPose": {
                          "pitch": -0.2,
                          "roll": -1.6,
                          "yaw": -26.3
                        },
                        "gender": "male",
                        "age": 25,
                        "facialHair": {
                          "moustache": 0.1,
                          "beard": 0.1,
                          "sideburns": 0.1
                        },
                        "glasses": "NoGlasses",
                        "emotion": {
                          "anger": 0,
                          "contempt": 0,
                          "disgust": 0,
                          "fear": 0,
                          "happiness": 1,
                          "neutral": 0,
                          "sadness": 0,
                          "surprise": 0
                        },
                        "blur": {
                          "blurLevel": "low",
                          "value": 0.15
                        },
                        "exposure": {
                          "exposureLevel": "goodExposure",
                          "value": 0.73
                        },
                        "noise": {
                          "noiseLevel": "low",
                          "value": 0
                        },
                        "makeup": {
                          "eyeMakeup": false,
                          "lipMakeup": false
                        },
                        "accessories": [],
                        "occlusion": {
                          "foreheadOccluded": false,
                          "eyeOccluded": false,
                          "mouthOccluded": false
                        },
                        "hair": {
                          "bald": 0.16,
                          "invisible": false,
                          "hairColor": [
                            {
                              "color": "black",
                              "confidence": 0.96
                            },
                            {
                              "color": "brown",
                              "confidence": 0.91
                            },
                            {
                              "color": "blond",
                              "confidence": 0.4
                            },
                            {
                              "color": "gray",
                              "confidence": 0.34
                            },
                            {
                              "color": "other",
                              "confidence": 0.2
                            },
                            {
                              "color": "red",
                              "confidence": 0.05
                            },
                            {
                              "color": "white",
                              "confidence": 0
                            }
                          ]
                        }
                      },
                      "recognitionModel": "recognition_03"
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "img"
                ],
                "properties": {
                  "img": {
                    "type": "file",
                    "format": "image/jpg",
                    "description": "Send image of the person",
                    "required": true
                  },
                  "location": {
                    "type": "boolean",
                    "description": "Send true if you want face location in the image: face pixel position in the image",
                    "required": false
                  },
                  "features": {
                    "type": "boolean",
                    "description": "Send true if you want face features in the image: headPose, glasses, facialHair, makeup, accessories, hair",
                    "required": false
                  },
                  "landmark": {
                    "type": "boolean",
                    "description": "Send true if you want face landmarks in the image",
                    "required": false
                  },
                  "emotions": {
                    "type": "boolean",
                    "description": "Send true if you want the emotions of the faces in the image: smile, emotion",
                    "required": false
                  },
                  "characterstics": {
                    "type": "boolean",
                    "description": "Send true if you want face characterstics in the image: age, gender",
                    "required": false
                  },
                  "noises": {
                    "type": "boolean",
                    "description": "Send true if you want noises on the faces in the image: occlusion, exposure, noise, blur",
                    "required": false
                  },
                  "azureId": {
                    "type": "string",
                    "description": "You can send your own azure key",
                    "required": false
                  }
                }
              },
              "example": [
                {
                  "key": "img",
                  "type": "file",
                  "src": "../testing_photos/tom_holland_1.jpg"
                },
                {
                  "key": "location",
                  "value": "true",
                  "type": "text",
                  "disabled": true
                }
              ]
            }
          }
        }
      }
    },
...
    "/v1/facelist/{faceListId}": {
      "parameters": [
        {
          "name": "faceListId",
          "in": "path",
          "description": "Face list Id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "summary": "Create Facelist",
        "operationId": "FaceListCreate",
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Face List",
        "operationId": "FaceListDelete",
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "summary": "Get Face List data",
        "operationId": "FaceListgetfacedata",
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "patch": {
        "summary": "Update Facelist meta data",
        "operationId": "FaceListUpdate",
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add face to Face list",
        "operationId": "FaceListAddFace",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "img",
                  "userData"
                ],
                "properties": {
                  "img": {
                    "type": "file",
                    "format": "image/jpg",
                    "description": "Send image of the person",
                    "required": true
                  },
                  "userData": {
                    "type": "string",
                    "example": "one two three"
                  }
                }
              },
              "example": [
                {
                  "key": "img",
                  "type": "file",
                  "src": "/C:/Users/Prashant/Documents/System Integration/Final Project/testing_photos/tom_holland_3.jpg"
                },
                {
                  "key": "userData",
                  "value": "one two three",
                  "type": "text"
                }
              ]
            }
          }
        }
      }
    },
    "/v1/facelist/{faceListId}/{persistedFaceId}": {
      "parameters": [
        {
          "name": "faceListId",
          "in": "path",
          "description": "Face list Id",
          "required": true,
          "schema": {
            "type": "string"
          }
        },{
          "name": "persistedFaceId",
          "in": "path",
          "description": "Persisted Face Id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "delete": {
        "summary": "Delete Face from List",
        "operationId": "FaceListFaceDelete",
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {},
            "content": {
              "application/json": {
                "schema":{
                  "properties": {
                    "name": "success",
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/largefacelist/{largeFaceListId}/faces": {
      "get": {
        "summary": "Get Faces inside large face list",
        "operationId": "LargeFaceListGet",
        "parameters": [{
          "$ref": "#/components/parameters/largeFaceListId"
        },{
          "$ref": "#/components/parameters/start"
        },{
          "$ref": "#/components/parameters/top"
        }],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "azureId":{
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "start": {
        "name": "start",
        "in": "query",
        "required": true,
        "style": "form",
        "schema": {
          "type": "string",
          "example": "first-large"
        }
      },
      "top": {
        "name": "top",
        "in": "query",
        "required": true,
        "style": "form",
        "schema": {
          "type": "number"
        }
      },
      "largeFaceListId": {
        "name": "largeFaceListId",
        "in": "path",
        "description": "Large Face list Id",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    }
  },
  "security": [],
...
}